\name{glamlasso-package}
\alias{glamlasso-package}
\alias{glamlasso}
\docType{package}
\title{
\packageTitle{glamlasso}
}
\description{
\packageDescription{glamlasso}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{glamlasso}
\packageIndices{glamlasso}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{glamlasso}

Maintainer: \packageMaintainer{glamlasso}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
